+++ html/js/drawing.js
@@ -175,6 +175,7 @@ function drawSpace(rawspaces) {
                                 }
                                 // console.log(dt);
                                 // in case of corrupted JSON we skip updating the page
+                                // noinspection Annotator
                                 if (/^-{0,1}\d+$/.test(dt)) {
                                     document.getElementById(allmeasurements[i].name).innerText = dt;
                                 }